PEP#: 268 TITLE: PEP#268 SSLClientCertificate Propagation
Description: Code change

+++ src/Pegasus/Common/CIMMessageDeserializer.cpp
@@ -737,7 +737,7 @@ void CIMMessageDeserializer::_deserializeUserCertificate(
             subjectName,
             issuerName,
             versionNumber,
-            serialNumber,
+            (long)serialNumber,
             notBefore,
             notAfter,
             depth,
